scripts/#requirementstxttxt.txt
#requirementstxttxt
mysql-connector-python
requests
textblob
python-dotenv